backend/requirements.txt
fastapi
uvicorn[standard]
apify-client
openai
pydantic
python-dotenv
sqlalchemy[asyncio]
alembic
asyncpg
httpx
opencv-python-headless
python-telegram-bot
